pingen-fixer-backend/requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
pandas==2.1.4
xlsxwriter==3.1.9
requests==2.31.0
streamlit==1.31.0
plotly==5.18.0
langchain==0.1.0
openai==1.3.0
flask-limiter==3.5.0
flask-cors==4.0.0
redis==5.0.1
Pillow==10.1.0
reportlab==4.0.8